pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Python task-manager on django"
authors = ["a.cheremushkin <a.ch3re359@yandex.ru>"]
readme = "README.md"
packages = [{ include = "task_manager" }]


[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
python-dotenv = "^0.21.0"
django-bootstrap4 = "^22.2"
gunicorn = "^20.1.0"
django-filter = "^22.1"
psycopg2 = "^2.9.5"
rollbar = "^0.16.3"

[tool.poetry.group.dev.dependencies]
loguru = "^0.6.0"
flake8 = "^6.0.0"
dj-database-url = "^0.5.0"
coverage = "^6.5.0"


[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
